Добавлена ф-ия получения актуального курса в store
